int -> int32

+++ Source/Applications/ASCompilerApp.cpp
@@ -59,7 +59,7 @@ int main(int argc, char** argv)
     FO_STACK_TRACE_ENTRY();
 
     try {
-        InitApp(argc, argv, AppInitFlags::DisableLogTags);
+        InitApp(numeric_cast<int32>(argc), argv, AppInitFlags::DisableLogTags);
 
         FO_RUNTIME_ASSERT(!App->Settings.BakeOutput.empty());
